Move BindersModule to the appropriate package
